pyproject.toml
[tool.poetry]
name = "repl-nix-discordbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_discordbot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
discord-py-interactions = "^5.12.1"
pytest = "^8.2.1"
requests = "^2.32.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
